requirements.txt
fastapi~=0.118.0
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
python-dotenv

pydantic-settings~=2.11.0